fix: add auto auth redirect back
