+++ studio/components/interfaces/SignIn/SignInForm.tsx
@@ -1,7 +1,7 @@
 import HCaptcha from '@hcaptcha/react-hcaptcha'
 import { useStore } from 'hooks'
 import { useParams } from 'hooks/misc/useParams'
-import { auth } from 'lib/gotrue'
+import { auth, getNextPath } from 'lib/gotrue'
 import Link from 'next/link'
 import { useRouter } from 'next/router'
 import { useRef, useState } from 'react'
@@ -17,7 +17,7 @@ const signInSchema = object({
 const SignInForm = () => {
   const { ui } = useStore()
   const router = useRouter()
-  const { returnTo } = useParams()
+  const { next: returnTo } = useParams()
   const { cache } = useSWRConfig()
 
   const [captchaToken, setCaptchaToken] = useState<string | null>(null)
@@ -52,7 +52,7 @@ const SignInForm = () => {
       // @ts-ignore
       cache.clear()
 
-      await router.push(returnTo ?? '/projects')
+      await router.push(getNextPath(returnTo))
     } else {
       setCaptchaToken(null)
       captchaRef.current?.resetCaptcha()
