+++ studio/components/interfaces/SignIn/SignInWithGitHub.tsx
@@ -1,7 +1,10 @@
-import { auth } from 'lib/gotrue'
+import { useParams } from 'hooks'
+import { auth, getNextPath } from 'lib/gotrue'
 import { Button, IconGitHub } from 'ui'
 
 const SignInWithGitHub = () => {
+  const { next: returnTo } = useParams()
+
   async function handleGithubSignIn() {
     try {
       const { error } = await auth.signInWithOAuth({
@@ -11,7 +14,7 @@ const SignInWithGitHub = () => {
             process.env.NEXT_PUBLIC_VERCEL_ENV === 'preview'
               ? process.env.NEXT_PUBLIC_VERCEL_URL
               : process.env.NEXT_PUBLIC_SITE_URL
-          }/projects`,
+          }${getNextPath(returnTo)}`,
         },
       })
       if (error) throw error
