+++ studio/components/layouts/SignInLayout/SignInLayout.tsx
@@ -1,14 +1,12 @@
-import { useParams, useStore } from 'hooks'
-import { auth, STORAGE_KEY } from 'lib/gotrue'
+import { useStore } from 'hooks'
+import { STORAGE_KEY } from 'lib/gotrue'
 import { observer } from 'mobx-react-lite'
 import Head from 'next/head'
 import Image from 'next/image'
 import Link from 'next/link'
-import { useRouter } from 'next/router'
 import { PropsWithChildren, useEffect, useState } from 'react'
 import { tweets } from 'shared-data'
-import { useSWRConfig } from 'swr'
-import { Button, IconFile, IconFileText } from 'ui'
+import { Button, IconFileText } from 'ui'
 
 type SignInLayoutProps = {
   heading: string
@@ -24,43 +22,8 @@ const SignInLayout = ({
   logoLinkToMarketingSite = false,
   children,
 }: PropsWithChildren<SignInLayoutProps>) => {
-  const { returnTo } = useParams()
-  const router = useRouter()
-  const { cache } = useSWRConfig()
   const { ui } = useStore()
   const { theme } = ui
-
-  useEffect(() => {
-    ;(async () => {
-      const { error } = await auth.initialize()
-
-      if (error) {
-        ui.setNotification({
-          category: 'error',
-          message: error.message,
-        })
-
-        return
-      }
-
-      const {
-        data: { session },
-      } = await auth.getSession()
-
-      if (session) {
-        ui.setNotification({
-          category: 'success',
-          message: `Signed in successfully!`,
-        })
-
-        // .clear() does actually exist on the cache object, but it's not in the types 🤦🏻
-        // @ts-ignore
-        cache.clear()
-
-        await router.push(returnTo ?? '/projects')
-      }
-    })()
-  }, [])
 
   const [quote, setQuote] = useState<{
     text: string
@@ -80,7 +43,15 @@ const SignInLayout = ({
       <Head>
         <script
           dangerouslySetInnerHTML={{
-            __html: `if (window.localStorage.getItem('${STORAGE_KEY}')) {window.location.replace('/projects')}`,
+            __html: `if (localStorage.getItem('${STORAGE_KEY}')) {
+              let returnTo = new URLSearchParams(location.search).get('next') ?? '/projects'
+
+              if (returnTo === 'new-project') {
+                returnTo = '/new/new-project'
+              }
+
+              location.replace(returnTo)
+            }`,
           }}
         />
       </Head>
